Move Decision and DMNElementReference tests to separate package
