+++ kie-dmn/kie-dmn-validation/src/main/java/org/kie/dmn/validation/DMNValidatorImpl.java
@@ -83,6 +83,7 @@ public class DMNValidatorImpl implements DMNValidator {
                 ks.getResources().newClassPathResource("dmn-validation-rules-bkm.drl", getClass() ),
                 ks.getResources().newClassPathResource("dmn-validation-rules-context.drl", getClass() ),
                 ks.getResources().newClassPathResource("dmn-validation-rules-decision.drl", getClass() ),
+                ks.getResources().newClassPathResource("dmn-validation-rules-dmnelementref.drl", getClass() ),
                 ks.getResources().newClassPathResource("dmn-validation-rules-dtable.drl", getClass() ),
                 ks.getResources().newClassPathResource("dmn-validation-rules-info-req.drl", getClass() ),
                 ks.getResources().newClassPathResource("dmn-validation-rules-input.drl", getClass() ),
