--- kie-dmn/kie-dmn-validation/src/test/resources/org/kie/dmn/validation/DECISION_MISSING_EXPR.dmn
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2016 Red Hat, Inc. and/or its affiliates.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<definitions id="DECISION_MISSING_EXPR" name="DECISION_MISSING_EXPR"
-    namespace="https://github.com/kiegroup/kie-dmn"
-    xmlns="http://www.omg.org/spec/DMN/20151101/dmn.xsd"
-    xmlns:feel="http://www.omg.org/spec/FEEL/20140401"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://www.omg.org/spec/DMN/20151101/dmn.xsd http://www.omg.org/spec/DMN/20151101/dmn.xsd ">
-    <decision name="Greeting Message" id="d_GreetingMessage">
-        <variable name="Greeting Message" typeRef="feel:string"/>
-        <informationRequirement>
-            <requiredInput href="#i_FullName"/>
-        </informationRequirement>
-    </decision>
-    <inputData name="Full Name" id="i_FullName">
-        <variable name="Full Name" typeRef="feel:string"/>
-    </inputData>
-</definitions>
