+++ kie-dmn/kie-dmn-validation/src/test/resources/org/kie/dmn/validation/decision/DECISION_MISSING_VAR.dmn
